SQLTraceListener: Retrieve calling application class and method
